+++ edsnlp/core/lazy_collection.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import contextlib
 import sys
+from functools import wraps
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -26,6 +27,14 @@ if TYPE_CHECKING:
     from edsnlp.data.base import BaseReader, BaseWriter
 
 INFER = type("INFER", (), {"__repr__": lambda self: "INFER"})()
+
+
+def with_non_default_args(fn: Callable) -> Callable:
+    @wraps(fn)
+    def wrapper(self, **kwargs):
+        return fn(self, **kwargs, _non_default_args=kwargs.keys())
+
+    return wrapper
 
 
 class MetaLazyCollection(type):
@@ -58,6 +67,26 @@ class LazyCollection(metaclass=MetaLazyCollection):
         return self.config.get("batch_size", 1)
 
     @property
+    def batch_by(self):
+        return self.config.get("batch_by", "docs")
+
+    @property
+    def sort_chunks(self):
+        return self.config.get("sort_chunks", False)
+
+    @property
+    def split_into_batches_after(self):
+        return self.config.get("split_into_batches_after")
+
+    @property
+    def chunk_size(self):
+        return self.config.get("chunk_size", self.config.get("batch_size", 128))
+
+    @property
+    def disable_implicit_parallelism(self):
+        return self.config.get("disable_implicit_parallelism", True)
+
+    @property
     def num_cpu_workers(self):
         return self.config.get("num_cpu_workers")
 
@@ -83,30 +112,60 @@ class LazyCollection(metaclass=MetaLazyCollection):
 
     @property
     def show_progress(self):
-        return self.config.get("show_progress")
+        return self.config.get("show_progress", False)
 
     @property
     def process_start_method(self):
         return self.config.get("process_start_method")
 
+    @with_non_default_args
     def set_processing(
         self,
-        batch_size: int = INFER,
-        num_cpu_workers: int = INFER,
-        num_gpu_workers: int = INFER,
-        backend: Literal["simple", "multiprocessing", "spark"] = INFER,
-        gpu_pipe_names: List[str] = INFER,
-        show_progress: bool = INFER,
-        process_start_method: bool = INFER,
-        gpu_worker_devices: List[str] = INFER,
-        cpu_worker_devices: List[str] = INFER,
+        batch_size: int = 1,
+        batch_by: Literal["docs", "words", "padded_words"] = "docs",
+        chunk_size: int = INFER,
+        sort_chunks: bool = False,
+        split_into_batches_after: str = INFER,
+        num_cpu_workers: Optional[int] = INFER,
+        num_gpu_workers: Optional[int] = INFER,
+        disable_implicit_parallelism: bool = True,
+        backend: Optional[Literal["simple", "multiprocessing", "spark"]] = INFER,
+        gpu_pipe_names: Optional[List[str]] = INFER,
+        show_progress: bool = False,
+        process_start_method: Optional[Literal["fork", "spawn"]] = INFER,
+        gpu_worker_devices: Optional[List[str]] = INFER,
+        cpu_worker_devices: Optional[List[str]] = INFER,
+        _non_default_args: Iterable[str] = (),
     ) -> "LazyCollection":
         """
         Parameters
         ----------
         batch_size: int
-            Number of documents to process at a time in a CPU/GPU worker (or in the
+            Number of documents to process at a time in a GPU worker (or in the
             main process if no workers are used).
+        batch_by: Literal["docs", "words", "padded_words"]
+            How to compute the batch size. Can be "docs" or "words" :
+
+            - "docs" (default) is the number of documents.
+            - "words" is the total number of words in the documents.
+            - "padded_words" is the total number of words in the documents, including
+               padding, assuming the documents are padded to the same length.
+        chunk_size: int
+            Number of documents to build before splitting into batches. Only used
+            with "simple" and "multiprocessing" backends. This is also the number of
+            documents that will be passed through the first components of the pipeline
+            until a GPU worker is used (then the chunks will be split according to the
+            `batch_size` and `batch_by` arguments).
+
+            By default, the chunk size is equal to the batch size, or 128 if the batch
+            size is not set.
+        sort_chunks: bool
+            Whether to sort the documents by size before splitting into batches.
+        split_into_batches_after: str
+            The name of the component after which to split the documents into batches.
+            Only used with "simple" and "multiprocessing" backends.
+            By default, the documents are split into batches as soon as the input
+            are converted into Doc objects.
         num_cpu_workers: int
             Number of CPU workers. A CPU worker handles the non deep-learning components
             and the preprocessing, collating and postprocessing of deep-learning
@@ -115,9 +174,13 @@ class LazyCollection(metaclass=MetaLazyCollection):
         num_gpu_workers: Optional[int]
             Number of GPU workers. A GPU worker handles the forward call of the
             deep-learning components. Only used with "multiprocessing" backend.
+        disable_implicit_parallelism: bool
+            Whether to disable OpenMP and Huggingface tokenizers implicit parallelism in
+            multiprocessing mode. Defaults to True.
         gpu_pipe_names: Optional[List[str]]
             List of pipe names to accelerate on a GPUWorker, defaults to all pipes
             that inherit from TorchComponent. Only used with "multiprocessing" backend.
+            Inferred from the pipeline if not set.
         backend: Optional[Literal["simple", "multiprocessing", "spark"]]
             The backend to use for parallel processing. If not set, the backend is
             automatically selected based on the input data and the number of workers.
@@ -131,9 +194,9 @@ class LazyCollection(metaclass=MetaLazyCollection):
         show_progress: Optional[bool]
             Whether to show progress bars (only applicable with "simple" and
             "multiprocessing" backends).
-        process_start_method: Optional[bool]
+        process_start_method: Optional[Literal["fork", "spawn"]]
             Whether to use "fork" or "spawn" as the start method for the multiprocessing
-            backend.
+            backend. The default is "fork" on Unix systems and "spawn" on Windows.
 
             - "fork" is the default start method on Unix systems and is the fastest
                 start method, but it is not available on Windows, can cause issues
@@ -151,8 +214,7 @@ class LazyCollection(metaclass=MetaLazyCollection):
         -------
         LazyCollection
         """
-        kwargs = dict(locals())
-        kwargs.pop("self")
+        kwargs = {k: v for k, v in locals().items() if k in _non_default_args}
         return LazyCollection(
             reader=self.reader,
             writer=self.writer,